Added UTF8 library and support for UO window title. Started rework of KeystrokeSender using a common base class and abstract methods.

+++ src/globals.cpp
@@ -21,14 +21,14 @@ std::vector<std::string> g_scriptFileList;
 
 ClientProfile*  getLoadedClientProfile()
 {
-    if ((g_loadedClientProfile >= 0) && (g_loadedClientProfile < g_clientProfiles.size()))
+    if ((g_loadedClientProfile >= 0) && (g_loadedClientProfile < (int)g_clientProfiles.size()))
         return &g_clientProfiles[g_loadedClientProfile];
     return nullptr;
 }
 
 ScriptsProfile* getLoadedScriptsProfile()
 {
-    if ((g_loadedScriptsProfile >= 0) && (g_loadedClientProfile < g_scriptsProfiles.size()))
+    if ((g_loadedScriptsProfile >= 0) && (g_loadedClientProfile < (int)g_scriptsProfiles.size()))
         return &g_scriptsProfiles[g_loadedScriptsProfile];
     return nullptr;
 }
